Added wasm compile target
So far the code actually run nicely in a browser.  The exceptions are
not handle in the C++.  For some reason, compiling with the flag
DISABLE_EXCEPTION_CATCHING=0 makes the browser crash when loading the
assembly.

cpudiag is stored in a C style array

```bash
xxd --include ../roms/cpudiag.bin > cpudiag.cpp
```